+++ digital-web/src/main/java/com/digital/domain/Entity.java
@@ -2,7 +2,7 @@ package com.digital.domain;
 
 public interface Entity {
 
-	public int getId();
-	public void setId(int id);
+	public Long getId();
+	public void setId(Long id);
 	
 }
